refactor(cleanup): consolidate error classes and fix global imports
