+++ packages/yukigo/src/utils/helpers.ts
@@ -1,12 +1,6 @@
 import { parseDocument } from "yaml";
 import { InspectionRule } from "../analyzer/index.js";
 
-export class UnexpectedNode extends Error {
-  constructor(nodeCons: string, context: string) {
-    super(`${nodeCons} not expected in ${context}.`)
-  }
-}
-
 type MulangInspection = {
   inspection: string;
   binding: string;
